database migration with checks

+++ database/migrations/2016_06_16_135109_create_jobs_table.php
@@ -12,17 +12,19 @@ class CreateJobsTable extends Migration
      */
     public function up()
     {
-        Schema::create('jobs', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('queue');
-            $table->longText('payload');
-            $table->tinyInteger('attempts')->unsigned();
-            $table->tinyInteger('reserved')->unsigned();
-            $table->unsignedInteger('reserved_at')->nullable();
-            $table->unsignedInteger('available_at');
-            $table->unsignedInteger('created_at');
-            $table->index(['queue', 'reserved', 'reserved_at']);
-        });
+        if (!Schema::hasTable('jobs')) {
+            Schema::create('jobs', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('queue');
+                $table->longText('payload');
+                $table->tinyInteger('attempts')->unsigned();
+                $table->tinyInteger('reserved')->unsigned();
+                $table->unsignedInteger('reserved_at')->nullable();
+                $table->unsignedInteger('available_at');
+                $table->unsignedInteger('created_at');
+                $table->index(['queue', 'reserved', 'reserved_at']);
+            });
+        }
     }
 
     /**
@@ -32,6 +34,8 @@ class CreateJobsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('jobs');
+        if (Schema::hasTable('jobs')) {
+            Schema::drop('jobs');
+        }
     }
 }
